+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # CHANGELOG
 
+#### July 2018
+
+- 03 Jul: Complete #43, add journal support for non-company currencies and exchange rate.
+
 #### June 2018
 
+- 29 Jun: In Journal add support for get, post, put of memo.
+- 29 Jun: Use ISO8601 for Journal dates.
 - 28 Jun: Fix #41 Exchange rate get (all) doesn't return an ISO8601 date.
 
 #### 26 June 2018 V2.4-1.7